Share libvirt participant test fixtures to clear Sonar duplication
Extract the deterministic participant-implementation manifest, selection,
action-result, and no-op libvirt driver into
tests/libvirt_participant_fixtures.py and import them from both the
acceptance tests and the proof driver, removing the near-identical copies
that tripped SonarCloud's new-code duplication gate.
