+++ implementations/python/tests/libvirt_participant_proof.py
@@ -8,23 +8,20 @@ invariants.
 
 This driver requires no live libvirt daemon; it is suitable for CI pipelines
 and conformance proofs. The ``DeterministicParticipantDomainAdapter`` is used
-throughout; live domain execution requires a custom adapter.
+throughout; live domain execution requires a custom adapter. The deterministic
+participant manifest/selection/action-result fixtures are shared with the
+acceptance tests in ``libvirt_participant_fixtures``.
 """
 
 from __future__ import annotations
 
-from collections.abc import Mapping
 from dataclasses import dataclass
 from pathlib import Path
 
 from aces_backend_libvirt.manifest import create_libvirt_manifest
 from aces_backend_libvirt.participant_runtime import LibvirtParticipantRuntime
 from aces_backend_libvirt.provisioner import LibvirtProvisioner
-from aces_contracts.contracts import (
-    ParticipantActionResultModel,
-    ParticipantImplementationManifestModel,
-    ParticipantImplementationSelectionModel,
-)
+from aces_contracts.contracts import ParticipantActionResultModel
 from aces_contracts.participant_binding import ParticipantActionAdmissionRequest
 from aces_processor.compiler import compile_runtime_model
 from aces_processor.models import (
@@ -36,66 +33,12 @@ from aces_processor.models import (
 from aces_runtime.control_plane import RuntimeControlPlane
 from aces_runtime.registry import RuntimeTarget
 from aces_sdl.parser import parse_sdl
-
-# ---------------------------------------------------------------------------
-# Proof identity constants
-# ---------------------------------------------------------------------------
-
-_PROOF_AGENT_NAME = "libvirt-deterministic-agent"
-_PROOF_AGENT_VERSION = "1.0.0"
-_PROOF_AGENT_IDENTITY = {"name": _PROOF_AGENT_NAME, "version": _PROOF_AGENT_VERSION}
-_PROOF_MANIFEST_REF = "contracts/fixtures/participant-implementation-manifest/libvirt-deterministic.json"
-_PROOF_MANIFEST_DIGEST = "sha256:" + "1" * 64
-_PROOF_POLICY_ID = "libvirt-paper-agent-policy"
-_PROOF_POLICY_VERSION = "1.0.0"
-_PROOF_POLICY_DIGEST = "sha256:" + "3" * 64
-
-_PROOF_WITHHELD_REFS = (
-    "content.evaluator-notes",
-    "nodes.customer-db.services.postgres",
-    "nodes.wazuh-manager",
-    "nodes.wazuh-indexer",
-    "nodes.participant-policy-gate",
+from libvirt_participant_fixtures import (
+    NullLibvirtDriver,
+    build_action_result,
+    build_implementation_manifest,
+    build_implementation_selection,
 )
-
-_PROOF_CONCEPT_BINDINGS = [
-    {"scope": "implementation_kind", "family": "apparatus-declarations"},
-    {"scope": "capabilities.supported_participant_contracts", "family": "apparatus-declarations"},
-    {"scope": "capabilities.supported_decision_surface_modes", "family": "apparatus-declarations"},
-    {"scope": "capabilities.tool_affordance_expectations", "family": "tools-and-artifacts"},
-    {"scope": "capabilities.exposure_policy_kinds", "family": "provenance-and-evidence"},
-]
-
-
-# ---------------------------------------------------------------------------
-# Null driver — no live libvirt connection needed for the structural proof
-# ---------------------------------------------------------------------------
-
-
-class _NullLibvirtDriver:
-    """No-op libvirt driver for the structural proof.
-
-    The proof never calls realize() or destroy(), so no real libvirt daemon
-    is needed. Returning empty results keeps the provisioner constructor happy.
-    """
-
-    def realize(self, *, networks, domains):
-        from aces_backend_libvirt.driver import DriverResult
-
-        return DriverResult()
-
-    def destroy(self, *, networks, domains):
-        from aces_backend_libvirt.driver import DriverResult
-
-        return DriverResult()
-
-    def realized_addresses(self):
-        return frozenset()
-
-
-# ---------------------------------------------------------------------------
-# Result dataclass
-# ---------------------------------------------------------------------------
 
 
 @dataclass(frozen=True)
@@ -107,82 +50,6 @@ class LibvirtParticipantProofResult:
     behavior_history_violations: tuple[tuple[str, str], ...] = ()
 
 
-# ---------------------------------------------------------------------------
-# Manifest and selection builders
-# ---------------------------------------------------------------------------
-
-
-def _build_proof_manifest() -> ParticipantImplementationManifestModel:
-    return ParticipantImplementationManifestModel.model_validate(
-        {
-            "schema_version": "participant-implementation-manifest/v1",
-            "identity": _PROOF_AGENT_IDENTITY,
-            "implementation_kind": "agent",
-            "supported_contract_versions": [
-                "participant-implementation-manifest-v1",
-                "participant-implementation-provenance-v1",
-                "participant-episode-state-envelope-v1",
-                "participant-episode-history-event-stream-v1",
-                "participant-behavior-history-event-stream-v1",
-            ],
-            "compatibility": {
-                "participant_runtimes": ["libvirt-qemu"],
-                "processors": ["aces-reference-processor"],
-                "backends": ["libvirt-qemu"],
-            },
-            "concept_bindings": _PROOF_CONCEPT_BINDINGS,
-            "constraints": {
-                "max_parallel_episodes": "1",
-                "simulation_disclosure": "deterministic-simulation: no live libvirt domain execution",
-            },
-            "capabilities": {
-                "supported_participant_contracts": [
-                    "participant-episode-state-envelope-v1",
-                    "participant-episode-history-event-stream-v1",
-                    "participant-behavior-history-event-stream-v1",
-                ],
-                "supported_decision_surface_modes": ["policy-directed"],
-                "tool_affordance_expectations": ["http-api"],
-                "exposure_policy_kinds": ["task-statement", "observation-stream"],
-            },
-        }
-    )
-
-
-def _build_proof_selection(
-    participant_address: str,
-    withheld_refs: tuple[str, ...],
-) -> ParticipantImplementationSelectionModel:
-    return ParticipantImplementationSelectionModel.model_validate(
-        {
-            "participant_address": participant_address,
-            "implementation_identity": _PROOF_AGENT_IDENTITY,
-            "manifest_ref": _PROOF_MANIFEST_REF,
-            "manifest_digest": _PROOF_MANIFEST_DIGEST,
-            "selected_decision_surface_mode": "policy-directed",
-            "participant_contract_versions": [
-                "participant-episode-state-envelope-v1",
-                "participant-behavior-history-event-stream-v1",
-            ],
-            "exposure_policy": {
-                "policy_id": _PROOF_POLICY_ID,
-                "policy_version": _PROOF_POLICY_VERSION,
-                "policy_digest": _PROOF_POLICY_DIGEST,
-                "exposure_policy_kinds": ["task-statement", "observation-stream"],
-                "disclosed_refs": [],
-                "withheld_refs": list(withheld_refs),
-                "tool_affordance_refs": [],
-                "visibility_scope_refs": [],
-            },
-        }
-    )
-
-
-# ---------------------------------------------------------------------------
-# Action result builder (generic, contracts-driven)
-# ---------------------------------------------------------------------------
-
-
 def _build_action_result(
     *,
     participant_address: str,
@@ -191,71 +58,21 @@ def _build_action_result(
     action_contract_address: str,
     contract: ParticipantActionContractRuntime,
 ) -> ParticipantActionResultModel | None:
-    """Build a valid succeeded action_result for the given compiled contract.
+    """Build a succeeded action_result, or None when the contract has no SEM-211 classes.
 
-    Reports all declared preconditions (with empty support_refs and
-    evidence_refs) to satisfy the SEM-211 completeness requirement. Reports
-    only ``no_effect`` effects, which require no target_refs or evidence_refs
-    and therefore cannot produce hidden-ref or boundary-evidence violations.
-
-    Returns ``None`` when the contract does not use SEM-211 action results.
+    Delegates the body to the shared ``build_action_result`` fixture; this wrapper
+    only adds the SEM-211 applicability gate, which depends on the compiled
+    contract object rather than the raw spec dict.
     """
     if not _contract_uses_sem211_action_results(contract):
         return None
-
-    preconditions_raw = contract.spec.get("preconditions", ())
-    effects_raw = contract.spec.get("effects", ())
-
-    preconditions = []
-    for pc in preconditions_raw:
-        if not isinstance(pc, Mapping) or not pc.get("precondition_id") or not pc.get("precondition_class"):
-            continue
-        preconditions.append(
-            {
-                "precondition_id": str(pc["precondition_id"]),
-                "precondition_class": str(pc["precondition_class"]),
-                "status": "satisfied",
-                "participant_address": participant_address,
-                "episode_id": episode_id,
-                "action_contract_address": action_contract_address,
-                "observation_point": f"{action_instance_id}:pc-{pc['precondition_id']}",
-                "support_refs": [],
-                "evidence_refs": [],
-            }
-        )
-
-    effects = []
-    for eff in effects_raw:
-        if not isinstance(eff, Mapping) or not eff.get("effect_id") or not eff.get("effect_class"):
-            continue
-        if str(eff["effect_class"]) == "no_effect":
-            effects.append(
-                {
-                    "effect_id": str(eff["effect_id"]),
-                    "effect_class": "no_effect",
-                    "description": str(eff.get("description", "No domain effect in deterministic proof.")),
-                }
-            )
-
-    return ParticipantActionResultModel.model_validate(
-        {
-            "status": "succeeded",
-            "participant_address": participant_address,
-            "episode_id": episode_id,
-            "action_instance_id": action_instance_id,
-            "action_contract_address": action_contract_address,
-            "observation_point": f"{action_instance_id}:terminal-observation",
-            "preconditions": preconditions,
-            "effects": effects,
-            "observations": [f"{action_instance_id}:terminal-observation"],
-            "evidence_refs": [],
-        }
+    return build_action_result(
+        participant_address=participant_address,
+        episode_id=episode_id,
+        action_instance_id=action_instance_id,
+        action_contract_address=action_contract_address,
+        contract_spec=contract.spec,
     )
-
-
-# ---------------------------------------------------------------------------
-# Public proof entry point
-# ---------------------------------------------------------------------------
 
 
 def run_libvirt_participant_proof(sdl_path: Path) -> LibvirtParticipantProofResult:
@@ -288,12 +105,12 @@ def run_libvirt_participant_proof(sdl_path: Path) -> LibvirtParticipantProofResu
     target = RuntimeTarget(
         name=manifest.name,
         manifest=manifest,
-        provisioner=LibvirtProvisioner(_NullLibvirtDriver()),
+        provisioner=LibvirtProvisioner(NullLibvirtDriver()),
         participant_runtime=participant_runtime,
     )
     control_plane = RuntimeControlPlane(target)
 
-    proof_manifest = _build_proof_manifest()
+    proof_manifest = build_implementation_manifest()
 
     errors: list[str] = []
 
@@ -355,7 +172,7 @@ def run_libvirt_participant_proof(sdl_path: Path) -> LibvirtParticipantProofResu
                 action_contract_address=action_address,
                 contract=contract,
             )
-            selection = _build_proof_selection(behavior_address, _PROOF_WITHHELD_REFS)
+            selection = build_implementation_selection(behavior_address)
 
             boundary_address = (
                 behavior.observation_boundary_addresses[0] if behavior.observation_boundary_addresses else None
